refactored get docs api call to be its own file for reusability
